+++ doc/kiro/README.md
@@ -1,0 +1,5 @@
+# These documents are generated from Kiro
+
+- product.md
+- structure.md
+- tech.md
