+++ doc/kiro/product.md
