+++ doc/kiro/structure.md
