+++ mcp-server/test-fixtures/README.md
@@ -1,0 +1,1 @@
+# Directory for TEST
