+++ mcp-server/type.ts
@@ -1,0 +1,7 @@
+/**
+ * mise.tomlから取得したtask
+ */
+export interface Task {
+  name: string;
+  description: string;
+}
